Python/requirements.txt
pylsl==1.12.2
serial==0.0.69
stomp.py==4.1.21
tobii-research==1.5.0
pynput==1.4
numpy==1.15.2
pandas==0.23.4
scikit-learn==0.20.0
requests==2.20.0
scipy==1.1.0
